TikTok/pyproject.toml
[project]
name = "tiktok-scraper"
version = "0.1.0"
description = "TikTok headless browser scraping evaluation — stealth testing, proxy validation, and detection analysis"
readme = "README.md"
authors = [
    { name = "Litus", email = "zegarraward@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "playwright>=1.52.0",
    "playwright-stealth-plugin>=2.5.0",
    "nodriver>=0.41",
    "httpx>=0.28.0",
    "selectolax>=0.3.0",
    "pydantic>=2.12.0",
    "python-dotenv>=1.2.0",
    "rich>=14.0.0",
]

[project.scripts]
tiktok-scraper = "tiktok_scraper:main"

[build-system]
requires = ["uv_build>=0.10.2,<0.11.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
    "integration: tests that make real network requests",
]

[dependency-groups]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=0.26.0",
    "ruff>=0.15.0",
]
